+++ exp_by_day/10.17/exp08a:5frame_physical.sh
@@ -1,4 +1,4 @@
 ## 5帧降采样，物理索引
 
 
-bash ./scripts/train.sh -d -F 5 -P -b 1 -c "0" -g 42 -w Exp08a
+bash ./scripts/train.sh -F 5 -P -b 4 -c "2,3,4,5,6,7" -g 42 -w Exp08a